PHPCAS-90 Empty Location headers are now ignored.
One proxy-authenticated web-service I was dealing with was setting an
invalid location header that I needed to remove on the web-service end.
Unfortunately, in PHP<5.3 it seems that there is no way to actually remove
a header, the best you can do is set it to an empty value. This fix prevents
phpCAS from trying to redirect to an empty URL.

git-svn-id: https://source.jasig.org/cas-clients/phpcas/branches/PHPCAS-90@22585 f5dbab47-78f9-eb45-b975-e544023573eb