PHPCAS-90 Standardized the CAS_Exception class hierarchy.
The CAS_Exception is now an interface rather than a class.

All exceptions thrown in phpCAS should implement this interface to allow them
to be caught as a category by clients. Each phpCAS exception should extend
an appropriate SPL exception class that best fits its type.

For example, an InvalidArgumentException in phpCAS should be defined as

		class CAS_InvalidArgumentException
			extends InvalidArgumentException
			implements CAS_Exception
		{ }

This definition allows the CAS_InvalidArgumentException to be caught as either
an InvalidArgumentException or as a CAS_Exception.

git-svn-id: https://source.jasig.org/cas-clients/phpcas/branches/PHPCAS-90@23514 f5dbab47-78f9-eb45-b975-e544023573eb